requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.0
sqlparse==0.5.1
uritemplate==4.1.1
